blame buffers using their native line ending
